refactor(Commands): added check for leaderboard channel incase server has deleted it

+++ idle-rpg/v2/idle-rpg/data/Commands.js
@@ -577,10 +577,12 @@ There's a command to get the invite link ${value}invite`);
     }
     const leaderboardChannel = await guild.channels.find(channel => channel.name === 'leaderboards' && channel.type === 'text' && channel.parent && channel.parent.name === 'Idle-RPG');
     const announcementChannel = await guild.channels.find(channel => channel.name === 'announcements' && channel.type === 'text' && channel.parent && channel.parent.name === 'Idle-RPG');
-    const leaderboardMessages = await leaderboardChannel.fetchMessages({ limit: 10 });
-    let resetMsg = '';
-    if (leaderboardChannel.size > 0 && leaderboardMessages.size > 0) {
-      await leaderboardMessages.array().forEach(msg => resetMsg = resetMsg.concat(`${msg.content}\n`) && msg.delete());
+    if (leaderboardChannel) {
+      const leaderboardMessages = await leaderboardChannel.fetchMessages({ limit: 10 });
+      let resetMsg = '';
+      if (leaderboardChannel.size > 0 && leaderboardMessages.size > 0) {
+        await leaderboardMessages.array().forEach(msg => resetMsg = resetMsg.concat(`${msg.content}\n`) && msg.delete());
+      }
     }
 
     const defaultConfig = {
